Terminal velocity multi-instance drifting.
More self-contained instances as we approach a windowing solution.
Currently each video instance has an associated playlist selector.
This may change in the future, but for now it's akin to a per-window
selector.

However, for sanity we're still destroying all previous instances before
spawning the next.
Also added rtv.offset for local sync offset on resync.