Add default guide sorting, based on file path.
Before, the guide would order the playlists in the order they were
cached (subject to network/browser).
This provides a bit more consistency, but is subject to sorting by the
file path.

It's not complicated, but picking a good sort method would still be
polarizing. Using playlist titles would ensure proper order but require
the names be consistent and also not trying to "game" the sorting.